Re-organize/move class definition files Adjust path references for moved class files
