+++ lib/classes/umw/outreach/residence.php
@@ -1,13 +1,15 @@
 <?php
 /**
- * Special treatment for the Areas of Study site
+ * Special treatment for the Residence Halls site
  */
-if ( ! class_exists( 'UMW_Outreach_Mods_Study' ) ) {
-	class UMW_Outreach_Mods_Study extends UMW_Outreach_Mods_Sub {
+namespace \UMW\Outreach\;
+
+if ( ! class_exists( 'Residence' ) ) {
+	class Residence extends Base {
 		/**
-		 * @var int $blog the ID of the Study blog
+		 * @var int $blog the ID of the Residence Life blog
 		 */
-		public $blog = 5;
+		public $blog = 30;
 		
 		function __construct() {
 			parent::__construct();
@@ -17,10 +19,10 @@ if ( ! class_exists( 'UMW_Outreach_Mods_Study' ) ) {
 			}
 			
 			/**
-			 * Fix the Areas of Study archives until I find a better way to handle this
+			 * Fix the Residence Hall archives until I find a better way to handle this
 			 */
-			add_action( 'template_redirect', array( $this, 'do_study_archives' ) );
-			add_action( 'genesis_before_loop', array( $this, 'do_program_feature' ), 11 );
+			/*add_action( 'template_redirect', array( $this, 'do_hall_archives' ) );*/
+			add_action( 'genesis_before_loop', array( $this, 'do_hall_feature' ), 11 );
 			
 			add_shortcode( 'wpv-oembed', array( $this, 'do_wpv_oembed' ) );
 		}
@@ -51,11 +53,11 @@ if ( ! class_exists( 'UMW_Outreach_Mods_Study' ) ) {
 		 * Insert the featured video or image above the content area
 		 * 		on an individual program
 		 */
-		function do_program_feature() {
-			if ( ! is_singular( 'areas' ) )
+		function do_hall_feature() {
+			if ( ! is_singular( 'residence-hall' ) )
 				return;
 			
-			$video = esc_url( get_post_meta( get_the_ID(), 'wpcf-video', true ) );
+			$video = esc_url( get_post_meta( get_the_ID(), 'wpcf-hall-video-url', true ) );
 			if ( empty( $video ) && ! has_post_thumbnail() ) {
 				return;
 			}
@@ -76,35 +78,28 @@ if ( ! class_exists( 'UMW_Outreach_Mods_Study' ) ) {
 		}
 		
 		/**
-		 * Fix the Areas of Study post type archives
+		 * Fix the Residence Halls post type archives
 		 */
 		function do_study_archives() {
 			if ( ! is_post_type_archive() && ! is_tax() )
 				return;
 			
-			if ( is_post_type_archive( 'areas' ) || is_tax( 'key' ) || is_tax( 'department' ) ) {
+			if ( is_post_type_archive( 'residence-hall' ) ) {
 				remove_action( 'genesis_loop', 'genesis_do_loop' );
-				add_action( 'genesis_loop', array( $this, 'do_study_loop' ) );
+				add_action( 'genesis_loop', array( $this, 'do_hall_loop' ) );
 			}
 		}
 		
 		/**
-		 * Output the Areas of Study A to Z list
+		 * Output the Residence Halls A to Z list
 		 */
-		function do_study_loop() {
+		function do_hall_loop() {
 			$args = array(
-				'post_type' => 'areas', 
+				'post_type' => 'residence-hall', 
 				'view'      => 106, 
 				'return_link' => 0, 
 				'alpha_links' => 0, 
 			);
-			if ( is_tax( 'key' ) || is_tax( 'department' ) ) {
-				$ob = get_queried_object();
-				if ( is_object( $ob ) && ! is_wp_error( $ob ) ) {
-					$args['tax_name'] = $ob->taxonomy;
-					$args['tax_term'] = $ob->slug;
-				}
-			}
 			$meat = '';
 			foreach ( $args as $k=>$v ) {
 				if ( is_numeric( $v ) )
@@ -115,30 +110,18 @@ if ( ! class_exists( 'UMW_Outreach_Mods_Study' ) ) {
 			$short = sprintf( '[atoz%s]', $meat );
 			$content = do_shortcode( $short );
 			
-			add_filter( 'genesis_post_title_text', array( $this, 'do_study_archive_title' ) );
+			add_filter( 'genesis_post_title_text', array( $this, 'do_hall_archive_title' ) );
 			add_filter( 'genesis_link_post_title', array( $this, '_return_false' ) );
 			$this->custom_genesis_loop( $content );
 			remove_filter( 'genesis_link_post_title', array( $this, '_return_false' ) );
-			remove_filter( 'genesis_post_title_text', array( $this, 'do_study_archive_title' ) );
+			remove_filter( 'genesis_post_title_text', array( $this, 'do_hall_archive_title' ) );
 		}
 		
 		/**
-		 * Return the title for the Areas of Study archive page
+		 * Return the title for the Residence Halls archive page
 		 */
-		function do_study_archive_title( $title ) {
-			if ( is_tax( 'key' ) ) {
-				$ob = get_queried_object();
-				if ( is_object( $ob ) && ! is_wp_error( $ob ) ) {
-					return __( sprintf( '%s A to Z', $ob->name ) );
-				}
-			} else if ( is_tax( 'department' ) ) {
-				$ob = get_queried_object();
-				if ( is_object( $ob ) && ! is_wp_error( $ob ) ) {
-					return __( sprintf( 'Areas of Study in %s', $ob->name ) );
-				}
-			}
-			
-			return __( 'Areas of Study A to Z' );
+		function do_hall_archive_title( $title ) {
+			return __( 'Residence Halls A to Z' );
 		}
 	}
 }
