+++ lib/classes/umw/outreach/study.php
@@ -1,13 +1,15 @@
 <?php
 /**
- * Special treatment for the Residence Halls site
+ * Special treatment for the Areas of Study site
  */
-if ( ! class_exists( 'UMW_Outreach_Mods_Residence' ) ) {
-	class UMW_Outreach_Mods_Residence extends UMW_Outreach_Mods_Sub {
+namespace \UMW\Outreach\;
+
+if ( ! class_exists( 'Study' ) ) {
+	class Study extends Base {
 		/**
-		 * @var int $blog the ID of the Residence Life blog
+		 * @var int $blog the ID of the Study blog
 		 */
-		public $blog = 30;
+		public $blog = 5;
 		
 		function __construct() {
 			parent::__construct();
@@ -17,10 +19,10 @@ if ( ! class_exists( 'UMW_Outreach_Mods_Residence' ) ) {
 			}
 			
 			/**
-			 * Fix the Residence Hall archives until I find a better way to handle this
+			 * Fix the Areas of Study archives until I find a better way to handle this
 			 */
-			/*add_action( 'template_redirect', array( $this, 'do_hall_archives' ) );*/
-			add_action( 'genesis_before_loop', array( $this, 'do_hall_feature' ), 11 );
+			add_action( 'template_redirect', array( $this, 'do_study_archives' ) );
+			add_action( 'genesis_before_loop', array( $this, 'do_program_feature' ), 11 );
 			
 			add_shortcode( 'wpv-oembed', array( $this, 'do_wpv_oembed' ) );
 		}
@@ -51,11 +53,11 @@ if ( ! class_exists( 'UMW_Outreach_Mods_Residence' ) ) {
 		 * Insert the featured video or image above the content area
 		 * 		on an individual program
 		 */
-		function do_hall_feature() {
-			if ( ! is_singular( 'residence-hall' ) )
+		function do_program_feature() {
+			if ( ! is_singular( 'areas' ) )
 				return;
 			
-			$video = esc_url( get_post_meta( get_the_ID(), 'wpcf-hall-video-url', true ) );
+			$video = esc_url( get_post_meta( get_the_ID(), 'wpcf-video', true ) );
 			if ( empty( $video ) && ! has_post_thumbnail() ) {
 				return;
 			}
@@ -76,28 +78,35 @@ if ( ! class_exists( 'UMW_Outreach_Mods_Residence' ) ) {
 		}
 		
 		/**
-		 * Fix the Residence Halls post type archives
+		 * Fix the Areas of Study post type archives
 		 */
 		function do_study_archives() {
 			if ( ! is_post_type_archive() && ! is_tax() )
 				return;
 			
-			if ( is_post_type_archive( 'residence-hall' ) ) {
+			if ( is_post_type_archive( 'areas' ) || is_tax( 'key' ) || is_tax( 'department' ) ) {
 				remove_action( 'genesis_loop', 'genesis_do_loop' );
-				add_action( 'genesis_loop', array( $this, 'do_hall_loop' ) );
+				add_action( 'genesis_loop', array( $this, 'do_study_loop' ) );
 			}
 		}
 		
 		/**
-		 * Output the Residence Halls A to Z list
+		 * Output the Areas of Study A to Z list
 		 */
-		function do_hall_loop() {
+		function do_study_loop() {
 			$args = array(
-				'post_type' => 'residence-hall', 
+				'post_type' => 'areas', 
 				'view'      => 106, 
 				'return_link' => 0, 
 				'alpha_links' => 0, 
 			);
+			if ( is_tax( 'key' ) || is_tax( 'department' ) ) {
+				$ob = get_queried_object();
+				if ( is_object( $ob ) && ! is_wp_error( $ob ) ) {
+					$args['tax_name'] = $ob->taxonomy;
+					$args['tax_term'] = $ob->slug;
+				}
+			}
 			$meat = '';
 			foreach ( $args as $k=>$v ) {
 				if ( is_numeric( $v ) )
@@ -108,18 +117,30 @@ if ( ! class_exists( 'UMW_Outreach_Mods_Residence' ) ) {
 			$short = sprintf( '[atoz%s]', $meat );
 			$content = do_shortcode( $short );
 			
-			add_filter( 'genesis_post_title_text', array( $this, 'do_hall_archive_title' ) );
+			add_filter( 'genesis_post_title_text', array( $this, 'do_study_archive_title' ) );
 			add_filter( 'genesis_link_post_title', array( $this, '_return_false' ) );
 			$this->custom_genesis_loop( $content );
 			remove_filter( 'genesis_link_post_title', array( $this, '_return_false' ) );
-			remove_filter( 'genesis_post_title_text', array( $this, 'do_hall_archive_title' ) );
+			remove_filter( 'genesis_post_title_text', array( $this, 'do_study_archive_title' ) );
 		}
 		
 		/**
-		 * Return the title for the Residence Halls archive page
+		 * Return the title for the Areas of Study archive page
 		 */
-		function do_hall_archive_title( $title ) {
-			return __( 'Residence Halls A to Z' );
+		function do_study_archive_title( $title ) {
+			if ( is_tax( 'key' ) ) {
+				$ob = get_queried_object();
+				if ( is_object( $ob ) && ! is_wp_error( $ob ) ) {
+					return __( sprintf( '%s A to Z', $ob->name ) );
+				}
+			} else if ( is_tax( 'department' ) ) {
+				$ob = get_queried_object();
+				if ( is_object( $ob ) && ! is_wp_error( $ob ) ) {
+					return __( sprintf( 'Areas of Study in %s', $ob->name ) );
+				}
+			}
+			
+			return __( 'Areas of Study A to Z' );
 		}
 	}
 }
